GSoC25_H/requirements.txt
pillow
pandas
numpy
scipy
SPARQLWrapper
wikipedia
stanza
dspy

#dev
pyright
streamlit
torch
tqdm
transformers
pytz
psutil
matplotlib
scikit-learn
python-crfsuite
jsonlines
toml
fastcoref
requests
nltk
graphviz
fairseq
marisa-trie
sklearn-crfsuite 
gdown # for downloading the models from google drive
genre
sentence-transformers
rdflib
googletrans
tensorboardX
sentencepiece